backend/requirements.txt
fastapi>=0.124.4
uvicorn>=0.38.0
sqlmodel>=0.0.27
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic-core>=2.41.5
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
mcp>=1.0.0
python-jose>=3.3.0